MAINT: Restructuring of applications folder / creation of examples folder
