+++ src/porepy/utils/examples_utils.py
